requirements.txt
pandas
requests
config
selenium
PIL
python-telegram-bot
google.cloud